errors: move UserRequestError to scylla crate
