+++ scylla/src/transport/connection.rs
@@ -3,7 +3,7 @@ use futures::{future::RemoteHandle, FutureExt};
 use scylla_cql::errors::{
     BrokenConnectionError, BrokenConnectionErrorKind, ConnectionError, ConnectionSetupRequestError,
     ConnectionSetupRequestErrorKind, CqlEventHandlingError, CqlRequestKind, RequestError,
-    ResponseParseError, TranslationError, UserRequestError,
+    ResponseParseError, TranslationError,
 };
 use scylla_cql::frame::frame_errors::CqlResponseParseError;
 use scylla_cql::frame::request::options::{self, Options};
@@ -34,7 +34,7 @@ use tokio_openssl::SslStream;
 pub(crate) use ssl_config::SslConfig;
 
 use crate::authentication::AuthenticatorProvider;
-use crate::errors::QueryError;
+use crate::errors::{QueryError, UserRequestError};
 use scylla_cql::frame::response::authenticate::Authenticate;
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::convert::TryFrom;
